+++ django_geoip/management/commands/geoip_update.py
@@ -1,17 +1,19 @@
 # -*- coding: utf-8 -*-
-from django.core.management.base import BaseCommand
-from optparse import make_option
 import logging
-from django_geoip.management.ipgeobase import IpGeobase
+from optparse import make_option
+
+from django.core.management.base import BaseCommand
+from ..ipgeobase import IpGeobase
 
 
 class Command(BaseCommand):
+
     help = 'Updates django-geoip data stored in db'
     option_list = BaseCommand.option_list + (
         make_option('--clear',
             action='store_true',
             default=False,
-            help=u"Clear tables prior import",
+            help="Clear tables prior import"
         ),
     )