+++ docs/changelog.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+0.3.0 (TBA)
+-----------
+* Added python 3 support (3.2 - 3.3)
+* ***IMPORTANT*** Minimum required django 1.4.2, use version 0.2.8, if you don't want to upgrade.
+* ``GeoLocationFascade`` alias removed
+
 
 0.2.8 (2012-12-10)
 ------------------